ChronoTimer now uses Channels.
I had to create a ChannelEvent class which allows the ChronoTimer to
determine which Channel fired the event.

If you connect a Sensor to a Channel, and the Channel is active, it will
receive signals from the Sensor that will start or stop ChronoTimer
depending on which Channel you connected it to.