Added fields to accommodate the group race type.
Added a RaceType enum that contains Individual, Group,
ParallelIndividual, and ParallelGroup race types.

The endRun and start methods now also work for group and inidividual
race types. Start takes an int as a parameter representing the lane to
start. The stop method will also take the same.

The private method didNotFinish is now deprecated. I'll remove it in a
later commit. It did nothing for us but add extra unnecessary code.