Implemented the ParallelGroup race type.
Added the cancel(int lane) method to ChronoTimer, which takes the lane
number to cancel. The racer cancelled is removed from the run, reset,
and placed back in the FRONT of the queue.

Added a remove(Racer racer) method to the Run class. This is called when
you cancel a racer.

Added cancel() to the Racer class, which just resets the racer's start
and end times.

Cleared out some unused imports.